Add InputContainer.js and InputContainer.text.js 추가
4강 리덕스 - 4주 차 과제 1
- InputContainer.js 추가.
- InputContainer.test.js 추가.
