+++ src/InputContainer.test.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+
+import { render, fireEvent } from '@testing-library/react';
+
+import { useDispatch, useSelector } from 'react-redux';
+
+import InputContainer from './InputContainer';
+
+jest.mock('react-redux');
+
+test('InputContainer', () => {
+  const dispatch = jest.fn();
+
+  useDispatch.mockImplementation(() => dispatch);
+
+  useSelector.mockImplementation((selector) => selector({
+    taskTitle: 'New Title',
+  }));
+
+  const { getByText, getByDisplayValue } = render((
+    <InputContainer />
+  ));
+
+  expect(getByText(/추가/)).not.toBeNull();
+  expect(getByDisplayValue(/New Title/)).not.toBeNull();
+
+  fireEvent.click(getByText(/추가/));
+  expect(dispatch).toBeCalledWith({ type: 'addTask' });
+});
